% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualize.R
\name{actualize}
\alias{actualize}
\title{Move the car}
\usage{
actualize(car, actions, t, walls, ...)
}
\arguments{
\item{car}{car list}

\item{actions}{actions list}

\item{t}{numeric : time epsilon}
}
\value{
car object updated.
}
\description{
Move the car
}
\examples{
actualize(set(reset()), list(acc=2, steer=-20), 0.1, walls5)
}
