\name{Angell}
\alias{Angell}
\docType{data}

\title{Moral Integration of American Cities}

\description{
The \code{Angell} data frame has 43 rows and 4 columns.
The observations are 43 U. S. cities around 1950.

}

\format{
  This data frame contains the following columns:
  \describe{
    \item{moral}{
      Moral Integration: Composite of crime rate and
      welfare expenditures.
      }
    \item{hetero}{
      Ethnic Heterogenity: From percentages of nonwhite and
      foreign-born white residents.
      }
    \item{mobility}{
      Geographic Mobility: From percentages of residents
      moving into and out of the city.
      }
    \item{region}{
      A factor with levels: 
      \code{E} Northeast;
      \code{MW} Midwest;
      \code{S} Southeast;
      \code{W} West.
      }
    }
  } 

\source{
 Angell, R. C. (1951) The moral integration of American Cities.
 \emph{American Journal of Sociology} \bold{57} (part 2), 1--140.
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
}

\usage{
Angell
}

\keyword{datasets}
