\name{Anscombe}
\alias{Anscombe}
\docType{data}

\title{U. S. State Public-School Expenditures} 

\description{
The \code{Anscombe} data frame has 51 rows and 4 columns.
The observations are the U. S. states plus Washington, D. C. in 1970.

}

\format{
  This data frame contains the following columns:
  \describe{
    \item{education}{
      Per-capita education expenditures, dollars.
      }
    \item{income}{
     Per-capita income, dollars. 
      }
    \item{young}{
      Proportion under 18, per 1000. 
      }
    \item{urban}{
      Proportion urban, per 1000. 
      }
    }
  }


\source{
  Anscombe, F. J. (1981)
  \emph{Computing in Statistical Science Through APL}. Springer-Verlag.
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
}

\usage{
Anscombe
}
\keyword{datasets}
