\name{Baumann}
\alias{Baumann}
\docType{data}
\title{Methods of Teaching Reading Comprehension}
\description{
The \code{Baumann} data frame has 66 rows and 6 columns.
The data are from an experimental study conducted by Baumann and Jones, as reported
by Moore and McCabe (1993). Students were randomly assigned to one of three
experimental groups.
}
\format{
This data frame contains the following columns:
\describe{
\item{group}{ 
Experimental group;
a factor with levels:
\code{Basal}, traditional method of teaching;
\code{DRTA}, an innovative method;
\code{Strat}, another innovative method.
}
\item{pretest.1}{
First pretest.
}
\item{pretest.2}{
Second pretest.
}
\item{post.test.1}{
First post-test.
}
\item{post.test.2}{
Second post-test.
}
\item{post.test.3}{
Third post-test.
}
}
}

\source{
  Moore, D. S. and McCabe, G. P. (1993)
  \emph{Introduction to the Practice of Statistics, Second Edition.}
  Freeman [pp. 794--795].
}
\usage{
Baumann
}
\keyword{datasets}
