\name{Blackmore}
\alias{Blackmore}
\docType{data}
\title{Exercise Histories of Eating-Disordered and Control Subjects}
\usage{Blackmore}
\description{
The \code{Blackmore} data frame has 945 rows and 4 columns.
Blackmore and Davis's data on exercise histories of 138 teenaged
girls hospitalized for eating disorders and 98 control subjects.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{subject}{a factor with subject id codes.}
    \item{age}{age in years.}
    \item{exercise}{hours per week of exercise.}
    \item{group}{a factor with levels:
        \code{control}, Control subjects;
        \code{patient}, Eating-disordered patients.
    }
  }
}

\source{
 Personal communication from Elizabeth Blackmore and Caroline Davis,
 York University.
}

\keyword{datasets}
