\name{Can.pop}
\alias{Can.pop}
\docType{data}

\title{Canadian Population Data}

\description{
The \code{Can.pop} data frame has 15 rows and 1 columns.
Decennial time-series of Canadian population, 1851--1991.

}

\format{
  This data frame contains the following columns:
  \describe{
    \item{year}{
        census year.
        }
    \item{population}{
      Population, in millions
      }
    }
  }

\source{
 Urquhart, M. C. and Buckley, K. A. H. (Eds.) (1965)
 \emph{Historical Statistics of Canada}. Macmillan [p. 1369].
 
 Canada (1994)
 \emph{Canada Year Book}. Statistics Canada [Table 3.2].
}

\references{
  Fox, J. (1997)
  \emph{Applied Regression, Linear Models, and Related Methods.} Sage.
}

\usage{
Can.pop
}
\keyword{datasets}
