\name{CanPop}
\alias{CanPop}
\docType{data}

\title{Canadian Population Data}

\description{
The \code{CanPop} data frame has 16 rows and 2 columns.
Decennial time-series of Canadian population, 1851--2001.

}

\format{
  This data frame contains the following columns:
  \describe{
    \item{year}{
        census year.
        }
    \item{population}{
      Population, in millions
      }
    }
  }

\source{
 Urquhart, M. C. and Buckley, K. A. H. (Eds.) (1965)
 \emph{Historical Statistics of Canada}. Macmillan, p. 1369.
 
 Canada (1994)
 \emph{Canada Year Book}. Statistics Canada, Table 3.2.

 Statistics Canada:
 \url{http://www12.statcan.ca/english/census01/products/standard/popdwell/Table-PR.cfm}.
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
}

\usage{
CanPop
}
\keyword{datasets}
