% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarR6.R
\name{Car}
\alias{Car}
\title{Car}
\description{
Car

Car
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Car$new()}}
\item \href{#method-actualize}{\code{Car$actualize()}}
\item \href{#method-print}{\code{Car$print()}}
\item \href{#method-place_wheels}{\code{Car$place_wheels()}}
\item \href{#method-place_outer}{\code{Car$place_outer()}}
\item \href{#method-place_radars}{\code{Car$place_radars()}}
\item \href{#method-place_radars_inside}{\code{Car$place_radars_inside()}}
\item \href{#method-intersect_radar}{\code{Car$intersect_radar()}}
\item \href{#method-radar_distance}{\code{Car$radar_distance()}}
\item \href{#method-put_on_track}{\code{Car$put_on_track()}}
\item \href{#method-see_car}{\code{Car$see_car()}}
\item \href{#method-see_radars}{\code{Car$see_radars()}}
\item \href{#method-see_track}{\code{Car$see_track()}}
\item \href{#method-see_car_track}{\code{Car$see_car_track()}}
\item \href{#method-see_car_track_inside}{\code{Car$see_car_track_inside()}}
\item \href{#method-see_inputs}{\code{Car$see_inputs()}}
\item \href{#method-see_line}{\code{Car$see_line()}}
\item \href{#method-move}{\code{Car$move()}}
\item \href{#method-physics_limit}{\code{Car$physics_limit()}}
\item \href{#method-clone}{\code{Car$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$new(x = 0, y = 0, orien = 90, range = 50, t = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric : placement initial du centre du cylindre arrière de l'auto.}

\item{\code{y}}{numeric : placement initial du centre du cylindre arrière de l'auto.}

\item{\code{orien}}{numeric : orientation initiale de l'auto.}

\item{\code{range}}{numeric positive : longueur des radars.}

\item{\code{t}}{numeric positive : temps entre chaque actualisation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-actualize"></a>}}
\if{latex}{\out{\hypertarget{method-actualize}{}}}
\subsection{Method \code{actualize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$actualize(initialize = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-place_wheels"></a>}}
\if{latex}{\out{\hypertarget{method-place_wheels}{}}}
\subsection{Method \code{place_wheels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$place_wheels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-place_outer"></a>}}
\if{latex}{\out{\hypertarget{method-place_outer}{}}}
\subsection{Method \code{place_outer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$place_outer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-place_radars"></a>}}
\if{latex}{\out{\hypertarget{method-place_radars}{}}}
\subsection{Method \code{place_radars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$place_radars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-place_radars_inside"></a>}}
\if{latex}{\out{\hypertarget{method-place_radars_inside}{}}}
\subsection{Method \code{place_radars_inside()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$place_radars_inside()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-intersect_radar"></a>}}
\if{latex}{\out{\hypertarget{method-intersect_radar}{}}}
\subsection{Method \code{intersect_radar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$intersect_radar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radar_distance"></a>}}
\if{latex}{\out{\hypertarget{method-radar_distance}{}}}
\subsection{Method \code{radar_distance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$radar_distance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_on_track"></a>}}
\if{latex}{\out{\hypertarget{method-put_on_track}{}}}
\subsection{Method \code{put_on_track()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$put_on_track(walls, finish)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-see_car"></a>}}
\if{latex}{\out{\hypertarget{method-see_car}{}}}
\subsection{Method \code{see_car()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$see_car()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-see_radars"></a>}}
\if{latex}{\out{\hypertarget{method-see_radars}{}}}
\subsection{Method \code{see_radars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$see_radars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-see_track"></a>}}
\if{latex}{\out{\hypertarget{method-see_track}{}}}
\subsection{Method \code{see_track()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$see_track()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-see_car_track"></a>}}
\if{latex}{\out{\hypertarget{method-see_car_track}{}}}
\subsection{Method \code{see_car_track()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$see_car_track()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-see_car_track_inside"></a>}}
\if{latex}{\out{\hypertarget{method-see_car_track_inside}{}}}
\subsection{Method \code{see_car_track_inside()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$see_car_track_inside()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-see_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-see_inputs}{}}}
\subsection{Method \code{see_inputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$see_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-see_line"></a>}}
\if{latex}{\out{\hypertarget{method-see_line}{}}}
\subsection{Method \code{see_line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$see_line(col_speed_or_acc = "speed")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move"></a>}}
\if{latex}{\out{\hypertarget{method-move}{}}}
\subsection{Method \code{move()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$move(acc, steer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-physics_limit"></a>}}
\if{latex}{\out{\hypertarget{method-physics_limit}{}}}
\subsection{Method \code{physics_limit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$physics_limit(acc, steer, t)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{numeric positive : temps entre chaque actualisation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Car$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
