\name{carHexsticker}
\alias{carHexsticker}
\title{
View the Official Hex Sticker for the car Package}
\description{
Open the official hex sticker for the car package in your browser}
\usage{
carHexsticker()
}

\value{
Used for its side effect of openning the hex sticker for the car package in your browser.
}

\author{
John Fox \email{jfox@mcmaster.ca}
}

\examples{
\dontrun{
carHexsticker()
}
}

\keyword{misc}
