\name{Chirot}
\alias{Chirot}
\docType{data}

\title{The 1907 Romanian Peasant Rebellion}

\description{
The \code{Chirot} data frame has 32 rows and 5 columns.
The observations are counties in Romania.

}

\format{
 This data frame contains the following columns:
 \describe{
   \item{intensity}{
     Intensity of the rebellion
     }
   \item{commerce}{
     Commercialization of agriculture
     }
   \item{tradition}{
     Traditionalism
     }
   \item{midpeasant}{
     Strength of middle peasantry
     }
   \item{inequality}{
     Inequality of land tenure
     }
   }
 }

\source{
 Chirot, D. and C. Ragin (1975)
 The market, tradition and peasant rebellion: The case of Romania.
 \emph{American Sociological Review} \bold{40}, 428--444 [Table 1].
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
}


\usage{
Chirot
}

\keyword{datasets}
