\name{Cowles}
\alias{Cowles}
\docType{data}
\title{Cowles and Davis's Data on Volunteering}
\usage{Cowles}
\description{
The \code{Cowles} data frame has 1421 rows and 4 columns.
These data come from a study of the personality determinants
of volunteering for psychological research.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{neuroticism}{scale from Eysenck personality inventory}
    \item{extraversion}{scale from Eysenck personality inventory}
    \item{sex}{a factor with levels: \code{female}; \code{male}}
    \item{volunteer}{volunteeing, a factor with levels: \code{no}; \code{yes}}
  }
}

\source{
 Cowles, M. and C. Davis (1987)
 The subject matter of psychology: Volunteers.
 \emph{British Journal of Social Psychology} \bold{26}, 97--102.
}
 

\keyword{datasets}
