\name{DavisThin}
\alias{DavisThin}
\docType{data}
\title{Davis's Data on Drive for Thinness}
\usage{DavisThin}
\description{
The \code{DavisThin} data frame has 191 rows and 7 columns.
This is part of a larger dataset for a study of eating disorders.
The seven variables in the data frame comprise a "drive for thinness"
scale, to be formed by summing the items.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{DT1}{a numeric vector}
    \item{DT2}{a numeric vector}
    \item{DT3}{a numeric vector}
    \item{DT4}{a numeric vector}
    \item{DT5}{a numeric vector}
    \item{DT6}{a numeric vector}
    \item{DT7}{a numeric vector}
  }
}

\source{
    Davis, C., G. Claridge, and D. Cerullo (1997) 
    Personality factors 
    predisposing to weight preoccupation: A continuum approach to the 
    association between eating disorders and personality disorders. 
    \emph{Journal of Psychiatric Research} \bold{31}, 467--480. [personal communication from the authors.]
}

\references{
  Fox, J. and Weisberg, S. (2011) 
  \emph{An R Companion to Applied Regression}, Second Edition, Sage.
}


\keyword{datasets}
