\name{Depredations}
\alias{Depredations}
\docType{data}
\title{
Minnesota Wolf Depredation Data
}
\description{
Wolf depredations of livestock on Minnesota farms, 1976-1998.
}
\usage{Depredations}
\format{
  A data frame with 434 observations on the following 5 variables.
  \describe{
    \item{\code{longitude}}{longitude of the farm}
    \item{\code{latitude}}{latitude of the farm}
    \item{\code{number}}{number of depredations 1976-1998}
    \item{\code{early}}{number of depredations 1991 or before}
    \item{\code{late}}{number of depredations 1992 or later}
  }
}

\references{
Fox, J. and Weisberg, S. (2011) 
\emph{An R Companion to Applied Regression}, Second Edition, Sage.
  
Harper, Elizabeth K. and Paul, William J. and Mech, L. David and Weisberg, Sanford (2008),
Effectiveness of Lethal, Directed Wolf-Depredation Control in Minnesota,
\emph{Journal of Wildlife Management}, 72, 3, 778-784.
\url{http://pinnacle.allenpress.com/doi/abs/10.2193/2007-273}
}

\keyword{datasets}
