% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward_done.R
\name{done}
\alias{done}
\title{Know if the car episode is done or not.}
\usage{
done(reward)
}
\arguments{
\item{reward}{numeric}
}
\value{
TRUE or FALSE
}
\description{
Know if the car episode is done or not.
}
\examples{
done(-1)
}
