\name{Florida}
\alias{Florida}
\docType{data}
\title{Florida County Voting}
\description{
The \code{Florida} data frame has 67 rows and 11 columns.
Vote by county in Florida for President in the 2000 election.
}
\format{
This data frame contains the following columns:
\describe{
\item{GORE}{
Number of votes for Gore
}
\item{BUSH}{
Number of votes for Bush.
}
\item{BUCHANAN}{
Number of votes for Buchanan.
}
\item{NADER}{
Number of votes for Nader.
}
\item{BROWNE}{
Number of votes for Browne (whoever that is).
}
\item{HAGELIN}{
Number of votes for Hagelin (whoever that is).
}
\item{HARRIS}{
Number of votes for Harris (whoever that is).
}
\item{MCREYNOLDS}{
Number of votes for McReynolds (whoever that is).
}
\item{MOOREHEAD}{
Number of votes for Moorehead (whoever that is).
}
\item{PHILLIPS}{
Number of votes for Phillips (whoever that is).
}
\item{Total}{
Total number of votes.
}
}
}

\source{
Adams, G. D. and Fastnow, C. F. (2000)
A note on the voting irregularities in Palm Beach, FL.
Formerly at \samp{http://madison.hss.cmu.edu/}, but no longer available there.
}
\usage{
Florida
}
\keyword{datasets}
