\name{Freedman}
\alias{Freedman}
\docType{data}
\title{Crowding and Crime in U. S. Metropolitan Areas}
\description{
The \code{Freedman} data frame has 110 rows and 4 columns.
The observations are U. S. metropolitan areas with 1968 populations
of 250,000 or more. There are some missing data.
}
\format{
This data frame contains the following columns:
\describe{
\item{population}{
Total 1968 population, 1000s.
}
\item{nonwhite}{
Percent nonwhite population, 1960.
}
\item{density}{
Population per square mile, 1968.
}
\item{crime}{
Crime rate per 100,000, 1969.
}
}
}

\source{
 United States (1970)
 \emph{Statistical Abstract of the United States}. 
 Bureau of the Census.
}

\references{
  Fox, J. and Weisberg, S. (2011) 
  \emph{An R Companion to Applied Regression}, Second Edition, Sage.
  
  Freedman, J. (1975)
  \emph{Crowding and Behavior.} Viking.
}

\usage{
Freedman
}

\keyword{datasets}
