\name{Ginzberg}
\alias{Ginzberg}
\docType{data}

\title{Data on Depression}

\description{
The \code{Ginzberg} data frame has 82 rows and 6 columns.
The data are for psychiatric patients hospitalized for depression.
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{simplicity}{
      Measures subject's need to see the world in black and white.
      }
    \item{fatalism}{
      Fatalism scale.
      }
    \item{depression}{
      Beck self-report depression scale.
      }
    \item{adjsimp}{
      Adjusted Simplicity: Simplicity adjusted (by regression) for other
      variables thought to influence depression.
      }
    \item{adjfatal}{
      Adjusted Fatalism.
      }
    \item{adjdep}{
      Adjusted Depression.
      }
    }
  }

\source{
 Personal communication from Georges Monette, Department of
 Mathematics and Statistics, York University, with the permission
 of the original investigator.
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
}

\usage{
Ginzberg
}

\keyword{datasets}
