% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{intersect_radar}
\alias{intersect_radar}
\title{Stop the radar line at the first wall}
\usage{
intersect_radar(radar, walls)
}
\arguments{
\item{radar}{LINESTRING (1 line only).}

\item{walls}{LINESTRING (can be many connected lines).}
}
\value{

}
\description{
Stop the radar line at the first wall
}
\examples{
intersect_radar(radar=sf::st_linestring(matrix(c(0, 10, 0, 0), ncol=2, byrow=TRUE)), walls=sf::st_linestring(matrix(c(1, 7, -1, 4, 1, 0), ncol=2, byrow=TRUE)))
intersect_radar(radar=sf::st_linestring(matrix(c(-2, 10, -1, 5.5), ncol=2, byrow=TRUE)), walls=sf::st_linestring(matrix(c(1, 7, -1, 4, 1, 0), ncol=2, byrow=TRUE)))
}
