% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualize.R
\name{limits}
\alias{limits}
\title{Limit acceleration and steering according to physics}
\usage{
limits(car, actions, t)
}
\arguments{
\item{car}{car list.}

\item{actions}{list of acc and steer.}

\item{t}{numeric time interval.}
}
\value{
list of acc and steer.
}
\description{
Limit acceleration and steering according to physics
}
\examples{
limits(set(reset()), list(acc=30, steer=22), 1)
}
