\name{Mandel}
\alias{Mandel}
\docType{data}

\title{Contrived Collinear Data}

\description{
 The \code{Mandel} data frame has 8 rows and 3 columns.

}

\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{
      first predictor.
      }
      \item{x2}{
      second predictor.
      }
      \item{y}{
      response.
      }
    }
  }

\source{
 Mandel, J. (1982)
 Use of the singular value decomposition in regression analysis.
 \emph{The American Statistician} \bold{36}, 15--24.
}

\references{
  Fox, J. (1997)
  \emph{Applied Regression, Linear Models, and Related Methods.} Sage.
}


\usage{
Mandel
}

\keyword{datasets}
