\name{Pike}
\alias{Pike}
\docType{data}
\title{Northern pike catch per unit effort}
\description{
 Catch per unit effort data for 16 Minnesota lakes
}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{cpue}{Estimated catch per unit effect}
    \item{secpue}{Estimated standard error of CPUE}
    \item{density}{Estimated fish density}
    \item{sedens}{Estimated standard error of Density}
  }
}

\source{
R. Pierce, Minnesota Dept. of Natural Resources
}
\references{Weisberg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Section 4.6.3.}
\usage{
Pike
}
\keyword{datasets}
