\name{Pottery}
\alias{Pottery}
\docType{data}
\title{Chemical Composition of Pottery}
\description{
 The data give the chemical composition of ancient pottery found at four sites in Great Britain. They 
 appear in Hand, et al. (1994), and are used to illustrate MANOVA in the SAS Manual. (Suggested by Michael
 Friendly.)
}
\usage{
Pottery
}
\format{
  A data frame with 26 observations on the following 6 variables.
  \describe{
    \item{\code{Site}}{a factor with levels \code{AshleyRails} \code{Caldicot} \code{IsleThorns} \code{Llanedyrn}}
    \item{\code{Al}}{Aluminum}
    \item{\code{Fe}}{Iron}
    \item{\code{Mg}}{Magnesium}
    \item{\code{Ca}}{Calcium}
    \item{\code{Na}}{Sodium}
  }
}

\source{
  Hand, D. J., Daly, F., Lunn, A. D., McConway, K. J., and E., O. (1994)
  \emph{A Handbook of Small Data Sets}. Chapman and Hall.
}

\examples{
Pottery
}
\keyword{datasets}
