\name{Quartet}
\alias{Quartet}
\docType{data}

\title{Four Regression Datasets}

\description{
The \code{Quartet} data frame has 11 rows and 5 columns.
These are contrived data.

}

\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{
      X-values for datasets 1--3.
      }
    \item{y1}{
      Y-values for dataset 1.
      }
    \item{y2}{
      Y-values for dataset 2.
      }
    \item{y3}{
      Y-values for dataset 3.
      }
    \item{x4}{
      X-values for dataset 4.
      }
    \item{y4}{
      Y-values for dataset 4.
      }
    }
  }

\source{
 Anscombe, F. J. (1973)
 Graphs in statistical analysis.
 \emph{American Statistician} \bold{27}, 17--21.
}

\usage{
Quartet
}
\keyword{datasets}
