% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{radar_distance}
\alias{radar_distance}
\title{Add distances to walls detected from radars to car}
\usage{
radar_distance(car, walls)
}
\arguments{
\item{car}{car list.}
}
\value{
car list.
}
\description{
Add distances to walls detected from radars to car
}
\examples{
radar_distance(set(reset()), walls)
}
