% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{reset}
\alias{reset}
\title{Reset car object}
\usage{
reset(x = 0, y = 0, orien = 90, range = 100)
}
\arguments{
\item{x}{numeric.}

\item{y}{numeric.}

\item{orien}{numeric.}

\item{range}{numeric.}
}
\value{
car list with initial parameters (center position, spped and orientation).
}
\description{
Reset car object
}
\examples{
reset()
}
