% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward_done.R
\name{reward}
\alias{reward}
\title{Get the observed reward accordingly to the state}
\usage{
reward(car, walls, finish, ...)
}
\arguments{
\item{car}{car list.}

\item{walls}{LINESTRING (can be many connected lines).}

\item{finish}{LINESTRING.}
}
\value{
reward as a numeric value.
}
\description{
Get the observed reward accordingly to the state
}
\examples{
reward(set(reset()),
walls=sf::st_linestring(matrix(c(40, 70, 10, 70, -3, 63, -10, 50, -10, -2, 10, -2, 10, 50, 40, 50), ncol=2, byrow=TRUE)),
finish=sf::st_linestring(matrix(c(40, 50, 40, 70), ncol=2, byrow=TRUE)))
}
