\name{Robey}
\alias{Robey}
\docType{data}

\title{Fertility and Contraception}

\description{
The \code{Robey} data frame has 50 rows and 3 columns.
The observations are developing nations around 1990.
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{region}{
      A factor with levels:
      \code{Africa};
      \code{Asia}, Asia and Pacific;
      \code{Latin.Amer}, Latin America and Caribbean;
      \code{Near.East}, Near East and North Africa.
      }
    \item{tfr}{
      Total fertility rate (children per woman).
      }
    \item{contraceptors}{
      Percent of contraceptors among married women of childbearing age.
      }
    }
  }

\source{
 Robey, B., Shea, M. A., Rutstein, O. and Morris, L. (1992) 
 The reproductive revolution: New survey findings. 
 \emph{Population Reports}. Technical Report M-11. 
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
}


\usage{
Robey
}
\keyword{datasets}
