% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{see_car}
\alias{see_car}
\title{See dimensions of a car}
\usage{
see_car(car, ...)
}
\arguments{
\item{car}{car list}
}
\value{
ggplot.
}
\description{
See dimensions of a car
}
\examples{
see_car(set(reset()))
}
