% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{see_car_track}
\alias{see_car_track}
\title{see the car on the track}
\usage{
see_car_track(car, walls, finish)
}
\arguments{
\item{car}{car list.}

\item{walls}{LINESTRING (can be many connected lines).}

\item{finish}{LINESTRING.}
}
\value{
ggplot.
}
\description{
see the car on the track
}
\examples{
see_car_track(set(reset()), walls, finish)
}
