% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{see_line}
\alias{see_line}
\title{See the car, the track and where the car has been}
\usage{
see_line(car, walls, finish, line, ...)
}
\arguments{
\item{car}{car list.}

\item{walls}{LINESTRING (can be many connected lines).}

\item{finish}{LINESTRING.}

\item{line}{list of every car in history.}

\item{...}{(optional).}
}
\value{
ggplot.
}
\description{
See the car, the track and where the car has been
}
\examples{

}
