% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{see_radar}
\alias{see_radar}
\title{See dimensions of radars on the car}
\usage{
see_radar(car)
}
\arguments{
\item{car}{car list}
}
\value{
ggplot.
}
\description{
See dimensions of radars on the car
}
\examples{
see_radar(set(reset()))
}
