% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{see_track}
\alias{see_track}
\title{See the racing track}
\usage{
see_track(walls, finish)
}
\arguments{
\item{walls}{LINESTRING (can be many connected lines).}

\item{finish}{LINESTRING.}
}
\value{
ggplot.
}
\description{
See the racing track
}
\examples{
see_track(sf::st_linestring(matrix(c(40, 70, 10, 70, -3, 63, -10, 50, -10, -1, 10, -1, 10, 50, 40, 50), ncol=2, byrow=TRUE)),
finish)
}
