% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{set}
\alias{set}
\title{Set every element to car (wheels, outer layer, radars, radars' distance).}
\usage{
set(car)
}
\arguments{
\item{car}{car list}
}
\value{
car list
}
\description{
Set every element to car (wheels, outer layer, radars, radars' distance).
}
\examples{
set(reset())
}
