\name{Transact}
\alias{Transact}
\docType{data}
\title{Transaction data}
\description{
Data on transaction times in branch offices of a large Australian bank.
}
\format{This data frame contains the following columns:
 \describe{
    \item{t1}{
 number of type 1 transactions
    }
    \item{t2}{
 number of type 2 transactions
    }
    \item{time}{
 total transaction time, minutes
    }
 }
}

\source{
Cunningham, R. and Heathcote, C.
(1989), Estimating a non-Gaussian regression model with multicollinearity.
Australian Journal of Statistics, 31,12-17.}
\references{
Fox, J. and Weisberg, S. (2011) 
\emph{An R Companion to Applied Regression}, Second Edition, Sage.

Weisberg, S. (2014) \emph{Applied Linear Regression}, Fourth Edition,  Wiley, Section 4.6.1.
}
\usage{
Transact
}
\keyword{datasets}
