\name{UN}
\alias{UN}
\docType{data}
\title{GDP and Infant Mortality}
\description{
The \code{UN} data frame has 207 rows and 2 columns.
The data are for 1998 and are from the United Nations; the observations
are nations of the world.
There are some missing data.
}
\format{
This data frame contains the following columns:
\describe{
\item{infant.mortality}{
Infant morality rate, infant deaths per 1000 live births.
}
\item{gdp}{
GDP per capita, in U.S.~dollars.
}
}
}

\source{
 United Nations (1998) 
 Social indicators.  
 Originally from \samp{http://www.un.org/Depts/unsd/social/main.htm} but no longer there.
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
  
  Fox, J. and Weisberg, S. (2011) 
  \emph{An R Companion to Applied Regression}, Second Edition, Sage.
}

\usage{
UN
}
\keyword{datasets}
