\name{US.pop}
\alias{US.pop}
\docType{data}

\title{Population of the United States}

\description{
  The \code{US.pop} data frame has 21 rows and 1 columns.
  This is a decennial time-series, from 1790 to 1990.
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{year}{
        census year.
        }
    \item{population}{
      Population in millions.
      }
    }
  }

\source{
 United States (1994)
 \emph{Statistical Abstract of the United States.}
 Bureau of the Census.
}

\references{
  Fox, J. (1997)
  \emph{Applied Regression, Linear Models, and Related Methods.} Sage.
}

\usage{
US.pop
}

\keyword{datasets}
