\name{Vocab}
\alias{Vocab}
\docType{data}

\title{Vocabulary and Education}

\description{
The \code{Vocab} data frame has 968 rows and 2 columns.
The observations are respondents to the 1989 U. S. General Social Survey.
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{education}{
      Education, in years.
      }
    \item{vocabulary}{
      Vocabulary test score: number correct on a 10-word test.
      }
    }
  }

\source{
 National Opinion Research Center (1989)
 \emph{General Social Survey.}
 Distributed by the Inter-University Consortium for Political and Social
 Research.
}

\references{
  Fox, J. (1997)
  \emph{Applied Regression, Linear Models, and Related Methods.} Sage.
}


\usage{
Vocab
}

\keyword{datasets}
