% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.R
\name{wheels}
\alias{wheels}
\title{Add wheels and outer layer to car}
\usage{
wheels(car)
}
\arguments{
\item{car}{car list.}
}
\value{
car list.
}
\description{
Add wheels and outer layer to car
}
\examples{
wheels(reset())
}
