\name{car-deprecated}
\alias{car-deprecated}
\alias{bootCase}
\alias{bootCase.lm}
\alias{bootCase.glm}
\alias{bootCase.nls}
\alias{nextBoot}
\alias{nextBoot.lm}
\alias{nextBoot.nls}

\title{Deprecated Functions in the car Package}
\description{
These functions are provided for compatibility with older versions of the \pkg{car} package only, and may be removed eventually. Commands that worked in versions of the \pkg{car} package prior to version 3.0-0 will not necessarily work in version 3.0-0 and beyond, or may not work in the same manner.
}
\usage{

bootCase(...)

nextBoot(...)

}
\arguments{
  \item{...}{arguments to pass to methods.}
}

\details{
These functions are replaced by \code{\link{Boot}}.
}


\seealso{See Also  \code{\link{Boot}}}



\keyword{regression}
