\name{car-internal.Rd}
\alias{car-internal.Rd}
\alias{.carEnv}
\title{
Internal Objects for the \pkg{car} package
}
\description{
These objects (currently only the \code{.carEnv} environment) are exported for
technical reasons and are not for direct use.
}

\author{John Fox \email{jfox@mcmaster.ca}}

\keyword{misc}

