\name{car-package}
\alias{car-package}
\alias{car}
\docType{package}
\title{
Companion to Applied Regression
}
\description{
Functions and Datasets to Accompany J. Fox and S. Weisberg, 
An R Companion to Applied Regression, Second Edition, Sage, 2011.
}
\details{
\tabular{ll}{
Package: \tab car\cr
Version: \tab 2.1-3\cr
Date: \tab 2016-07-23\cr
Depends: \tab R (>= 3.2.0)\cr
Imports: \tab MASS, mgcv, nnet, pbkrtest (>= 0.3-2), quantreg, grDevices, utils, stats, graphics\cr
Suggests: \tab  alr4, boot, leaps, lme4, lmtest, Matrix, MatrixModels, mgcv, nlme, rgl (>= 0.93.960), sandwich, SparseM, survival, survey\cr
License: \tab GPL (>= 2)\cr
URL: \tab \url{http://CRAN.R-project.org/package=car}, 
	\url{http://socserv.socsci.mcmaster.ca/jfox/Books/Companion},
	\url{https://r-forge.r-project.org/projects/car/} \cr
}
}


\author{
John Fox <jfox@mcmaster.ca> and Sanford Weisberg. We are grateful to Douglas Bates, Gabriel Baud-Bovy,
David Firth, Michael Friendly, Gregor Gorjanc, Spencer Graves, 
Richard Heiberger, Rafael Laboissiere, Georges Monette, Henric Nilsson, Derek Ogle, Brian Ripley, 
Achim Zeleis, and R Core for various suggestions and contributions.

Maintainer: John Fox <jfox@mcmaster.ca>
}
\keyword{ package }
