\name{car-package}
\alias{car-package}
\alias{car}
\docType{package}
\title{
Companion to Applied Regression
}
\description{
This package accompanies Fox, J. and Weisberg, S., \emph{An R Companion to Applied Regression},
Second Edition, Sage, 2011.
}
\details{
\tabular{ll}{
Package: \tab car\cr
Version: \tab 2.0-15\cr
Date: \tab 2012/09/30\cr
Depends: \tab R (>= 2.1.1), stats, graphics, MASS, nnet\cr
Suggests: \tab  alr3, leaps, lme4, lmtest, nlme, sandwich, mgcv, pbkrtest, rgl, survival, survey\cr
License: \tab GPL (>= 2)\cr
URL: \tab \url{http://CRAN.R-project.org/package=car}, 
	\url{http://socserv.socsci.mcmaster.ca/jfox/Books/Companion},
	\url{https://r-forge.r-project.org/projects/car/} \cr
}
}


\author{
John Fox <jfox@mcmaster.ca> and Sanford Weisberg. We are grateful to Douglas Bates,
David Firth, Michael Friendly, Gregor Gorjanc, Spencer Graves, 
Richard Heiberger, Rafael Laboissiere, Georges Monette, Henric Nilsson, Derek Ogle, Brian Ripley, 
Achim Zeleis, and R Core for various suggestions and contributions.

Maintainer: John Fox <jfox@mcmaster.ca>
}
\keyword{ package }
