\name{car-package}
\alias{car-package}
\alias{car}
\docType{package}
\title{
Companion to Applied Regression
}
\description{
This package accompanies J. Fox and S. Weisberg, An R Companion to Applied Regression,
Second Edition, Sage, forthcoming.
}
\details{
\tabular{ll}{
Package: \tab car\cr
Version: \tab 2.0-0\cr
Date: \tab 2010/03/16\cr
Depends: \tab R (>= 2.1.1), stats, graphics, MASS, nnet, leaps, survival, art\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://www.r-project.org, http://socserv.socsci.mcmaster.ca/jfox/\cr
}
}


\author{
John Fox <jfox@mcmaster.ca> and Sanford Weisberg. We are grateful to Douglas Bates,
David Firth, Michael Friendly, Gregor Gorjanc, Spencer Graves, 
Richard Heiberger, Georges Monette, Henric Nilsson, Brian Ripley, 
and Achim Zeleis for various suggestions and contributions.

Maintainer: John Fox <jfox@mcmaster.ca>
}
\keyword{ package }
