\name{carWeb}
\alias{carWeb}

\title{ Access to the R Companion to Applied Regression Website }
\description{
  This function will access the website for \emph{An R Companion to Applied Regression}.}

\usage{
carWeb(...)
}
\arguments{
  \item{...}{optional arguments if present are passed to utils::browseURL
 }
}

\value{
  Displays the web page for \emph{An R Companion to Applied Regression}, Third Edition, Sage 
}

\keyword{ interface }
