\name{Davis}
\alias{Davis}
\docType{data}

\title{Self-Reports of Height and Weight}

\description{
The \code{Davis} data frame has 200 rows and 5 columns.
The subjects were men and women engaged in regular exercise.
There are some missing data.
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{sex}{
      A factor with levels:
      \code{F}, female;
      \code{M}, male.
      }
    \item{weight}{
      Measured weight in kg. 
      }
    \item{height}{
      Measured height in cm. 
      }
    \item{repwt}{
      Reported weight in kg. 
      }
    \item{repht}{
      Reported height in cm. 
      }
    }
  }


\source{
 Personal communication from C. Davis, Departments of
 Physical Education and Psychology, York University.
}

\references{
  Davis, C. (1990)
  Body image and weight preoccupation: A comparison between exercising
  and non-exercising women.
  \emph{Appetite}, \bold{15}, 13--21.
  
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
  
  Fox, J. and Weisberg, S. (2011) 
  \emph{An R Companion to Applied Regression}, Second Edition, Sage.
}


\usage{
Davis
}

\keyword{datasets}
