\name{Sahlins}
\alias{Sahlins}
\docType{data}

\title{Agricultural Production in Mazulu Village}

\description{
The \code{Sahlins} data frame has 20 rows and 2 columns.
The observations are households in a Central African village.
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{consumers}{
      Consumers/Gardener, ratio of consumers to productive individuals.
      }
    \item{acres}{
      Acres/Gardener, amount of land cultivated per gardener.
      }
    }
  }

\source{
 Sahlins, M. (1972)
 \emph{Stone Age Economics.} 
 Aldine
 [Table 3.1]. 
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
}


\usage{
Sahlins
}

\keyword{datasets}
