\name{USPop}
\alias{USPop}
\docType{data}

\title{Population of the United States}

\description{
  The \code{USPop} data frame has 22 rows and 1 columns.
  This is a decennial time-series, from 1790 to 2000.
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{year}{
        census year.
        }
    \item{population}{
      Population in millions.
      }
    }
  }

\source{
U.S.~Census Bureau: \url{http://www.census-charts.com/Population/pop-us-1790-2000.html}, downloaded 1 May 2008.
}

\references{
  Fox, J. (2008)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Second Edition. Sage.  
}

\usage{
USPop
}

\keyword{datasets}
